course crud

+++ app/Http/Controllers/CourseController.php
@@ -73,6 +73,11 @@ class CourseController extends Controller
      */
     public function destroy(Course $course, Request $request)
     {
+        if ($course->has()->subcourse->exists()) {
+            $request->session()->flash('error', 'Course can not be deleted because it has one or more subcourses.');
+            return redirect()->route('courses.index');
+        }
+
         $course->delete();
         $request->session()->flash('message', 'Course has been deleted successfully.');
